add special bbox handling
